moved enums inside related models, updated validation
