+++ src/main/java/academy/softserve/eschool/model/Lesson.java
@@ -5,10 +5,12 @@ import java.util.HashSet;
 import java.util.Set;
 
 import javax.persistence.*;
+import javax.validation.constraints.FutureOrPresent;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.Size;
 
+import academy.softserve.eschool.model.Mark.MarkType;
 import lombok.*;
 
 @Entity
@@ -28,6 +30,7 @@ public class Lesson {
 	private byte lessonNumber;
 	@NotNull
 	@Temporal(TemporalType.DATE)
+	@FutureOrPresent
 	private Date date;
 	@Size(max=500)
 	private String hometask;
