+++ src/main/java/academy/softserve/eschool/model/Mark.java
@@ -43,4 +43,8 @@ public class Mark {
 		this.student = student;
 		this.lesson = lesson;
 	}
+	
+	public enum MarkType {
+		Control, Practic, Module, Labaratorna
+	}
 }
