--- src/main/java/academy/softserve/eschool/model/MarkType.java
@@ -1,5 +1,0 @@
-package academy.softserve.eschool.model;
-
-public enum MarkType {
-		Control, Practic, Module, Labaratorna
-}
