--- src/main/java/academy/softserve/eschool/model/Role.java
@@ -1,5 +1,0 @@
-package academy.softserve.eschool.model;
-
-public enum Role {
-	ROLE_TEACHER, ROLE_USER, ROLE_ADMIN
-}
