--- src/main/java/academy/softserve/eschool/model/Sex.java
@@ -1,5 +1,0 @@
-package academy.softserve.eschool.model;
-
-public enum Sex {
-	male, female
-}
