+++ src/main/java/academy/softserve/eschool/model/User.java
@@ -102,4 +102,15 @@ public class User {
 		this.avatar = avatar;
 		this.description = description;
 	}
+	
+	public enum Role {
+		ROLE_TEACHER, ROLE_USER, ROLE_ADMIN
+	}
+
+	public enum Sex {
+		male, female
+	}
 }
+
+
+
