+++ src/main/java/academy/softserve/eschool/service/StudentService.java
@@ -1,26 +1,26 @@
 package academy.softserve.eschool.service;
 
-import academy.softserve.eschool.dto.EditUserDTO;
-import academy.softserve.eschool.dto.NYTransitionDTO;
-import academy.softserve.eschool.dto.StudentDTO;
-import academy.softserve.eschool.model.Clazz;
-import academy.softserve.eschool.model.Role;
-import academy.softserve.eschool.model.Student;
-import academy.softserve.eschool.model.User;
-import academy.softserve.eschool.repository.ClassRepository;
-import academy.softserve.eschool.repository.StudentRepository;
-import academy.softserve.eschool.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Service;
+import static academy.softserve.eschool.auxiliary.LoginGeneratorController.transliteration;
+import static academy.softserve.eschool.auxiliary.PasswordGenerator.generatePassword;
 
-import javax.management.loading.ClassLoaderRepository;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static academy.softserve.eschool.auxiliary.LoginGeneratorController.transliteration;
-import static academy.softserve.eschool.auxiliary.PasswordGenerator.generatePassword;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Service;
+
+import academy.softserve.eschool.dto.EditUserDTO;
+import academy.softserve.eschool.dto.NYTransitionDTO;
+import academy.softserve.eschool.dto.StudentDTO;
+import academy.softserve.eschool.model.Clazz;
+import academy.softserve.eschool.model.Student;
+import academy.softserve.eschool.model.User;
+import academy.softserve.eschool.model.User.Role;
+import academy.softserve.eschool.repository.ClassRepository;
+import academy.softserve.eschool.repository.StudentRepository;
+import academy.softserve.eschool.repository.UserRepository;
 
 @Service
 public class StudentService {
